move logger to /pkg directory
